quint/src/itf.ts
/*
 * Support for the Informal Trace Format (ITF):
 * https://apalache.informal.systems/docs/adr/015adr-trace.html
 *
 * Igor Konnov, Informal Systems, 2023
 *
 * Copyright (c) Informal Systems 2021. All rights reserved.
 * Licensed under the Apache 2.0.
 * See License.txt in the project root for license information.
 */

import { Either, left, right, merge } from '@sweet-monads/either'
import { zip } from 'lodash'

import { QuintEx } from './quintIr'

// The minimal value that can be reliably represented with number
const minJsInt = -(2n ** 53n) + 1n
// The maximal value that can be reliably represented with number
const maxJsInt = (2n ** 53n) - 1n

/**
 * Convert a typed Quint expression into an object that matches the JSON
 * representation of the ITF trace.
 *
 * @param vars variable names
 * @param trace a sequence of expressions that represent the states
 * @returns an object that represent the trace in the ITF format
 */
export function toItf(vars: string[], trace: QuintEx): Either<string, any> {
  if (trace.kind !== 'app' || trace.opcode !== 'List') {
    return left(`Expected a list of records, found: ${trace.kind}`)
  }

  const exprToItf = (ex: QuintEx): Either<string, any> => {
    switch (ex.kind) {
      case 'int':
        if (ex.value >= minJsInt && ex.value <= maxJsInt) {
          // OK to convert to a number, when saving to JSON
          return right(ex.value)
        } else {
          // convert to a special structure, when saving to JSON
          return right({ '#bigint': `${ex.value}` })
        }

      case 'str':
      case 'bool':
        return right(ex.value)

      case 'app':
        switch (ex.opcode) {
          case 'List':
            return merge(ex.args.map(exprToItf))

          case 'Set':
            return merge(ex.args.map(exprToItf))
                  .mapRight(es => { return { '#set': es } })

          case 'Tup':
            return merge(ex.args.map(exprToItf))
                  .mapRight(es => { return { '#tup': es } })

          case 'Rec': {
            if (ex.args.length % 2 !== 0) {
              return left('record: expected an even number of arguments, found:' + ex.args.length)
            }
            return merge(ex.args.map(exprToItf)).mapRight(kvs => {
              const keys = kvs.filter((k, i) => i % 2 == 0)
              const values = kvs.filter((k, i) => i % 2 == 1)
              let obj: any = {}
              zip(keys, values).forEach(([k, v]) => {
                obj[k] = v
              })
              return obj
            })
          }

          case 'Map':
            return merge(ex.args.map(exprToItf)).mapRight(pairs => {
              return { '#map': pairs.map(p => p['#tup']) }
            })

          default:
            return left(`Unexpected operator type: ${ex.opcode}`)
        }

      default:
        return left(`Unexpected expression kind: ${ex.kind}`)
    }
  }

  return exprToItf(trace).mapRight(states => {
    return {
      "vars": vars,
      "states": states,
    }
  })
}
